# customer/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /customer/estimate-price/{origin}/{destination}:
        get:
            tags:
                - Customer
            operationId: Customer_EstimatePrice
            parameters:
                - name: origin
                  in: path
                  required: true
                  schema:
                    type: string
                - name: destination
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.customer.GetEstimatePriceReply'
    /customer/get-verify-code/{Telephone}:
        get:
            tags:
                - Customer
            description: |-
                rpc CreateCustomer (CreateCustomerRequest) returns (CreateCustomerReply);
                	rpc UpdateCustomer (UpdateCustomerRequest) returns (UpdateCustomerReply);
                	rpc DeleteCustomer (DeleteCustomerRequest) returns (DeleteCustomerReply);
            operationId: Customer_GetCustomer
            parameters:
                - name: Telephone
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.customer.GetCustomerReply'
    /customer/login:
        post:
            tags:
                - Customer
            description: 登录
            operationId: Customer_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.customer.LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.customer.LoginRes'
    /customer/logout:
        post:
            tags:
                - Customer
            description: 登出
            operationId: Customer_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.customer.LogoutReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.customer.LogoutRes'
    /helloworld/{name}:
        get:
            tags:
                - Greeter
            description: Sends a greeting
            operationId: Greeter_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/helloworld.v1.HelloReply'
components:
    schemas:
        api.customer.GetCustomerReply:
            type: object
            properties:
                Code:
                    type: integer
                    format: int32
                Message:
                    type: string
                verifyCode:
                    type: string
                verifyCodeTime:
                    type: string
                verifyCodeLife:
                    type: string
        api.customer.GetEstimatePriceReply:
            type: object
            properties:
                origin:
                    type: string
                destination:
                    type: string
                price:
                    type: string
        api.customer.LoginReq:
            type: object
            properties:
                telephone:
                    type: string
                verifyCode:
                    type: string
        api.customer.LoginRes:
            type: object
            properties:
                code:
                    type: integer
                    format: int32
                message:
                    type: string
                token:
                    type: string
                tokenCreateAt:
                    type: string
                tokenLifeTime:
                    type: string
        api.customer.LogoutReq:
            type: object
            properties: {}
        api.customer.LogoutRes:
            type: object
            properties:
                Code:
                    type: integer
                    format: int32
                Message:
                    type: string
        helloworld.v1.HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
tags:
    - name: Customer
    - name: Greeter
      description: The greeting service definition.